requirements.txt
asgiref==3.8.1
Django==5.0.7
dnspython==2.6.1
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
